pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "starpulse"
version = "0.1.0"
description = "Daily GitHub star growth detector"
requires-python = ">=3.11"
dependencies = [
  "google-cloud-bigquery>=3.25.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
cache_dir = "/private/tmp/starpulse-pytest-cache"
